refactor: extract transfer copy helper
